+++ robotidy/cli.py
@@ -46,7 +46,7 @@ click.rich_click.OPTION_GROUPS = {
             "name": "Global formatting settings",
             "options": ["--spacecount", "--line-length", "--lineseparator", "--separator", "--startline", "--endline"],
         },
-        {"name": "File exclusion", "options": ["--exclude", "--extend-exclude"]},
+        {"name": "File exclusion", "options": ["--exclude", "--extend-exclude", "--ignore-gitignore"]},
         {
             "name": "Other",
             "options": ["--target-version", "--verbose", "--color", "--output", "--version", "--help"],
@@ -267,6 +267,9 @@ def print_transformers_list(target_version: int):
     ),
 )
 @click.option(
+    "--ignore-gitignore", is_flag=True, show_default=True, help="Do not skip files listed in **.gitignore** files."
+)
+@click.option(
     "--config",
     type=click.Path(
         exists=True,
@@ -409,6 +412,7 @@ def cli(
     src: Tuple[str, ...],
     exclude: Optional[Pattern],
     extend_exclude: Optional[Pattern],
+    ignore_gitignore: bool,
     overwrite: bool,
     diff: bool,
     color: bool,
@@ -478,6 +482,7 @@ def cli(
         src=src,
         exclude=exclude,
         extend_exclude=extend_exclude,
+        ignore_gitignore=ignore_gitignore,
         overwrite=overwrite,
         show_diff=diff,
         formatting_config=formatting_config,